requirements.txt
googlesearch-python
requests
beautifulsoup4
Scrapy
